+++ src/formatters/html/TylerHtmlCompressor.php
@@ -31,6 +31,12 @@ class TylerHtmlCompressor extends Component implements IFormatter
     public $noComments = true;
 
     /**
+     * The maximum number of rows that the formatter runs on
+     * @var int
+     */
+    public $maxNumberRows = 50000;
+
+    /**
      * @param string $html
      * @return string
      */
@@ -41,10 +47,15 @@ class TylerHtmlCompressor extends Component implements IFormatter
             'extra' => $this->extra,
         ];
 
-        \Yii::beginProfile('countHtmlStrings');
+        \Yii::beginProfile('countHtmlRows');
             $count = substr_count($html, "\n") + 1;
-            \Yii::info('Number of HTML strings: ' . $count);
-        \Yii::endProfile('countHtmlStrings');
+            \Yii::trace('Number of HTML rows: ' . $count);
+            if ($count > $this->maxNumberRows) {
+                \Yii::info("Not run: " . self::class . ". Too many lines: {$count}. Can be no more than: {$this->maxNumberRows}");
+                return $html;
+            }
+
+        \Yii::endProfile('countHtmlRows');
 
         $result = HtmlCompressor::compress((string) $html, $options);